requirements.txt
beautifulsoup4==4.12.3
faiss-gpu==1.7.2
fastapi==0.110.0
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-openai==0.0.7
langchainhub==0.1.14
langserve==0.0.43
langsmith==0.1.9
lxml==5.1.0
openai==1.12.0
python-dotenv==1.0.1
selenium==4.18.1
sse-starlette==2.0.0
starlette==0.36.3
uvicorn==0.27.1